ft[dependencies]: Makes distinction between mujoco sensor and pybullet ones to avoid confusion and to allow the usage of mujoco without installation of pybullet.
