+++ urdfenvs/sensors/mujoco/fsd_sensor.py
@@ -1,0 +1,4 @@
+from urdfenvs.sensors.fsd_sensor import FSDSensor
+
+class FSDSensorMujoco(FSDSensor):
+    pass
