+++ urdfenvs/sensors/mujoco/lidar.py
@@ -1,0 +1,8 @@
+import numpy as np
+
+from urdfenvs.sensors.lidar import Lidar
+
+
+class LidarMujoco(Lidar):
+    def __str__(self):
+        pass
