+++ urdfenvs/sensors/mujoco/occupancy_sensor.py
@@ -1,0 +1,4 @@
+from urdfenvs.sensors.occupancy_sensor import OccupancySensor
+
+class OccupancySensorMujocor(OccupancySensor):
+    pass
